requirements.txt
nose==1.3.0
boto==2.10.0

# For the vulnweb
Django==1.5.2
requests==1.2.3
django-celery==3.0.21

# We need to connect to the RDS database and create a low-privileged user
# sudo apt-get install python-mysqldb
